devmode

+++ level/7/start.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+if [ ! -f .dev ]; then
+	exit 1
+fi
 
 rm -rf ./raum
 cp -r "$lernbashpath/level/7/dungeon/raum/" ./raum
